weld guard: key on FLOAT32, the precision the exported file actually carries
The guard keyed on exact double equality, which cleaned up the demo l-bracket at
resolution 48 completely — and then the maintainer's own part at resolution 64
still exported 37 798 distinct float32 positions from 37 800 double-distinct
vertices on rung 052. Two pairs stayed apart in double and MET when written.

Binary STL stores float32 and every consumer re-welds by position, so a
difference below a float32 ULP does not survive the file. The canonical riser
case is exact in double, which is why the first cut worked at all; float32 is
not a safety margin over it but the correct bound — exactly the set of pairs the
FILE cannot tell apart.

Pairs that were ALREADY float32-coincident in the input are left alone: not this
operation's doing, reverting would not fix them, and treating them as collisions
would make the guard fight a pre-existing degeneracy every pass without
converging.

Also fixes this task's own design.bin reader, which was carried over from PR
305's gate table and read the version from the wrong offset — it does not fail
cleanly on a misalignment, it silently produces garbage, so the block size is
now asserted against the grid.

+++ core/tests/unit/test_cad_project.cpp
@@ -198,12 +198,17 @@ void weld_guard_positive_control(const Fixture& fx) {
     m.triangles.push_back({2 * i + 1, 2 * (i + 1) + 1, 2 * (i + 2)});
   }
 
+  // ★ KEYED AT FLOAT32 — the precision binary STL stores and therefore the
+  // precision at which two vertices become one. A double-precision key passed
+  // on the maintainer's part at resolution 64 while the exported file still
+  // welded two pairs; see cad_project.cpp's weld guard.
   auto coincident_positions = [](const TriangleMesh& mm) {
-    std::map<std::array<double, 3>, std::size_t> at;
+    std::map<std::array<float, 3>, std::size_t> at;
     std::size_t n = 0;
     for (std::size_t i = 0; i < mm.vertices.size(); ++i) {
-      const std::array<double, 3> k{mm.vertices[i].x, mm.vertices[i].y,
-                                    mm.vertices[i].z};
+      const std::array<float, 3> k{static_cast<float>(mm.vertices[i].x),
+                                   static_cast<float>(mm.vertices[i].y),
+                                   static_cast<float>(mm.vertices[i].z)};
       if (!at.emplace(k, i).second) ++n;
     }
     return n;
@@ -473,11 +478,12 @@ int main() {
   //
   // So the invariant has to be stated at the level the file is written at.
   {
-    std::map<std::array<double, 3>, std::size_t> at;
+    std::map<std::array<float, 3>, std::size_t> at;
     std::size_t coincident = 0;
     for (std::size_t i = 0; i < proj.vertices.size(); ++i) {
-      const std::array<double, 3> k{proj.vertices[i].x, proj.vertices[i].y,
-                                    proj.vertices[i].z};
+      const std::array<float, 3> k{static_cast<float>(proj.vertices[i].x),
+                                   static_cast<float>(proj.vertices[i].y),
+                                   static_cast<float>(proj.vertices[i].z)};
       if (!at.emplace(k, i).second) ++coincident;
     }
     std::printf("weld guard: %zu collisions found, %zu projected vertices put "
@@ -489,12 +495,12 @@ int main() {
           "the exported file welds by position and would stop being watertight");
     // The SOURCE mesh must have no coincident vertices either, or the check
     // above would be measuring marching cubes rather than the projection.
-    std::map<std::array<double, 3>, std::size_t> at0;
+    std::map<std::array<float, 3>, std::size_t> at0;
     std::size_t coincident0 = 0;
     for (std::size_t i = 0; i < exported.vertices.size(); ++i) {
-      const std::array<double, 3> k{exported.vertices[i].x,
-                                    exported.vertices[i].y,
-                                    exported.vertices[i].z};
+      const std::array<float, 3> k{static_cast<float>(exported.vertices[i].x),
+                                   static_cast<float>(exported.vertices[i].y),
+                                   static_cast<float>(exported.vertices[i].z)};
       if (!at0.emplace(k, i).second) ++coincident0;
     }
     CHECK(coincident0 == 0,
